+++ apps/backend/src/ranges/enums/index.ts
@@ -1,3 +1,2 @@
 export * from './action-type.enum';
 export * from './player-position.enum';
-export * from './position.enum';